requirements.txt
aiogram
python-dotenv
openai>=1.0.0
